Update capture-information.mdx
I suggest adding default value to the documentation. 
I rushed my life trying to understand what I thought was a bug, but finally no. That was only default setup that I found doing my tests.

In Boolean Capture Card
--------------------------------
The Retry is set to -> 2 by default
The Cancel option is set to -> On by default
And there is a default Cancel Confirmation message set to -> Are you sure you want to cancel?